Fix numerical overflows in NativeDataAccess

+++ com.oracle.truffle.r.runtime/src/com/oracle/truffle/r/runtime/data/NativeDataAccess.java
@@ -477,56 +477,56 @@ public final class NativeDataAccess {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return UnsafeAdapter.UNSAFE.getInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE);
+        return UnsafeAdapter.UNSAFE.getInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE);
     }
 
     public static double getDoubleNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return UnsafeAdapter.UNSAFE.getDouble(address + index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+        return UnsafeAdapter.UNSAFE.getDouble(address + (long) index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
     }
 
     public static byte getLogicalNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return RRuntime.int2logical(UnsafeAdapter.UNSAFE.getInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE));
+        return RRuntime.int2logical(UnsafeAdapter.UNSAFE.getInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE));
     }
 
     public static byte getRawNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return UnsafeAdapter.UNSAFE.getByte(address + index * Unsafe.ARRAY_BYTE_INDEX_SCALE);
+        return UnsafeAdapter.UNSAFE.getByte(address + (long) index * Unsafe.ARRAY_BYTE_INDEX_SCALE);
     }
 
     public static RComplex getComplexNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return RComplex.valueOf(UnsafeAdapter.UNSAFE.getDouble(address + index * 2 * Unsafe.ARRAY_DOUBLE_INDEX_SCALE),
-                        UnsafeAdapter.UNSAFE.getDouble(address + (index * 2 + 1) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE));
+        return RComplex.valueOf(UnsafeAdapter.UNSAFE.getDouble(address + index * 2L * Unsafe.ARRAY_DOUBLE_INDEX_SCALE),
+                        UnsafeAdapter.UNSAFE.getDouble(address + (index * 2L + 1L) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE));
     }
 
     public static double getComplexNativeMirrorDataR(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return UnsafeAdapter.UNSAFE.getDouble(address + index * 2 * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+        return UnsafeAdapter.UNSAFE.getDouble(address + index * 2L * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
     }
 
     public static double getComplexNativeMirrorDataI(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        return UnsafeAdapter.UNSAFE.getDouble(address + (index * 2 + 1) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+        return UnsafeAdapter.UNSAFE.getDouble(address + (index * 2L + 1L) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
     }
 
     public static CharSXPWrapper getStringNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
-        long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + index * Long.BYTES);
+        long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + (long) index * Long.BYTES);
         assert elemAddr != 0L;
         return (CharSXPWrapper) NativeDataAccess.lookup(elemAddr);
     }
@@ -534,7 +534,7 @@ public final class NativeDataAccess {
     public static Object getListElementNativeMirrorData(Object nativeMirror, int index) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
-        long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + index * Long.BYTES);
+        long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + (long) index * Long.BYTES);
         assert elemAddr != 0L;
         return NativeDataAccess.lookup(elemAddr);
     }
@@ -543,42 +543,42 @@ public final class NativeDataAccess {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putDouble(address + index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
+        UnsafeAdapter.UNSAFE.putDouble(address + (long) index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
     }
 
     public static void setNativeMirrorComplexRealPartData(Object nativeMirror, int index, double value) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putDouble(address + 2 * index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
+        UnsafeAdapter.UNSAFE.putDouble(address + 2L * index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
     }
 
     public static void setNativeMirrorComplexImaginaryPartData(Object nativeMirror, int index, double value) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putDouble(address + (2 * index + 1) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
+        UnsafeAdapter.UNSAFE.putDouble(address + (2L * index + 1L) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
     }
 
     public static void setNativeMirrorRawData(Object nativeMirror, int index, byte value) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putByte(address + index * Unsafe.ARRAY_BYTE_INDEX_SCALE, value);
+        UnsafeAdapter.UNSAFE.putByte(address + (long) index * Unsafe.ARRAY_BYTE_INDEX_SCALE, value);
     }
 
     public static void setNativeMirrorIntData(Object nativeMirror, int index, int value) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE, value);
+        UnsafeAdapter.UNSAFE.putInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE, value);
     }
 
     public static void setNativeMirrorLogicalData(Object nativeMirror, int index, byte logical) {
         long address = ((NativeMirror) nativeMirror).dataAddress;
         assert address != 0;
         assert index < ((NativeMirror) nativeMirror).length;
-        UnsafeAdapter.UNSAFE.putInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE, RRuntime.logical2int(logical));
+        UnsafeAdapter.UNSAFE.putInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE, RRuntime.logical2int(logical));
     }
 
     public static void setNativeMirrorStringData(Object nativeMirror, int index, CharSXPWrapper value) {
@@ -587,7 +587,7 @@ public final class NativeDataAccess {
         assert index < ((NativeMirror) nativeMirror).length;
 
         long asPointer = getPointer(value);
-        UnsafeAdapter.UNSAFE.putLong(address + index * Long.BYTES, asPointer);
+        UnsafeAdapter.UNSAFE.putLong(address + (long) index * Long.BYTES, asPointer);
     }
 
     public static void setNativeMirrorListData(Object nativeMirror, int index, Object value) {
@@ -597,7 +597,7 @@ public final class NativeDataAccess {
 
         if (value instanceof RBaseObject) {
             long asPointer = getPointer((RBaseObject) value);
-            UnsafeAdapter.UNSAFE.putLong(address + index * Long.BYTES, asPointer);
+            UnsafeAdapter.UNSAFE.putLong(address + (long) index * Long.BYTES, asPointer);
         } else {
             throw RInternalError.shouldNotReachHere();
         }
@@ -608,7 +608,7 @@ public final class NativeDataAccess {
         long address = mirror.dataAddress;
         assert address != 0;
         double[] data = new double[(int) mirror.length];
-        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_DOUBLE_BASE_OFFSET, data.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_DOUBLE_BASE_OFFSET, (long) data.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
         return data;
     }
 
@@ -617,7 +617,7 @@ public final class NativeDataAccess {
         long address = mirror.dataAddress;
         assert address != 0;
         double[] data = new double[(int) (mirror.length << 1)];
-        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_DOUBLE_BASE_OFFSET, data.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_DOUBLE_BASE_OFFSET, (long) data.length * Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
         return data;
     }
 
@@ -626,7 +626,7 @@ public final class NativeDataAccess {
         long address = mirror.dataAddress;
         assert address != 0;
         int[] data = new int[(int) mirror.length];
-        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_INT_BASE_OFFSET, data.length * Unsafe.ARRAY_INT_INDEX_SCALE);
+        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_INT_BASE_OFFSET, (long) data.length * Unsafe.ARRAY_INT_INDEX_SCALE);
         return data;
     }
 
@@ -636,7 +636,7 @@ public final class NativeDataAccess {
         long address = mirror.dataAddress;
         assert address != 0;
         byte[] data = new byte[(int) mirror.length];
-        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_BYTE_BASE_OFFSET, data.length * Unsafe.ARRAY_BYTE_INDEX_SCALE);
+        UnsafeAdapter.UNSAFE.copyMemory(null, address, data, Unsafe.ARRAY_BYTE_BASE_OFFSET, (long) data.length * Unsafe.ARRAY_BYTE_INDEX_SCALE);
         return data;
     }
 
@@ -646,7 +646,7 @@ public final class NativeDataAccess {
         assert address != 0;
         String[] data = new String[(int) mirror.length];
         for (int i = 0; i < mirror.length; i++) {
-            long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + i * Long.BYTES);
+            long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + (long) i * Long.BYTES);
             assert elemAddr != 0L;
             Object elem = lookup(elemAddr);
             assert elem instanceof CharSXPWrapper;
@@ -661,7 +661,7 @@ public final class NativeDataAccess {
         assert address != 0;
         Object[] data = new Object[(int) mirror.length];
         for (int i = 0; i < mirror.length; i++) {
-            long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + i * Long.BYTES);
+            long elemAddr = UnsafeAdapter.UNSAFE.getLong(address + (long) i * Long.BYTES);
             assert elemAddr != 0L;
             Object elem = lookup(elemAddr);
             data[i] = elem;
@@ -727,7 +727,7 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            UnsafeAdapter.UNSAFE.putInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE, value);
+            UnsafeAdapter.UNSAFE.putInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE, value);
         }
     }
 
@@ -737,7 +737,7 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            return RRuntime.int2logical(UnsafeAdapter.UNSAFE.getInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE));
+            return RRuntime.int2logical(UnsafeAdapter.UNSAFE.getInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE));
         }
     }
 
@@ -755,7 +755,7 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            UnsafeAdapter.UNSAFE.putInt(address + index * Unsafe.ARRAY_INT_INDEX_SCALE, RRuntime.logical2int(value));
+            UnsafeAdapter.UNSAFE.putInt(address + (long) index * Unsafe.ARRAY_INT_INDEX_SCALE, RRuntime.logical2int(value));
         }
     }
 
@@ -823,7 +823,7 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            UnsafeAdapter.UNSAFE.putInt(address + index * Unsafe.ARRAY_BYTE_INDEX_SCALE, value);
+            UnsafeAdapter.UNSAFE.putInt(address + (long) index * Unsafe.ARRAY_BYTE_INDEX_SCALE, value);
         }
     }
 
@@ -940,8 +940,8 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            UnsafeAdapter.UNSAFE.putDouble(address + index * 2 * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, re);
-            UnsafeAdapter.UNSAFE.putDouble(address + (index * 2 + 1) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, im);
+            UnsafeAdapter.UNSAFE.putDouble(address + index * 2L * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, re);
+            UnsafeAdapter.UNSAFE.putDouble(address + (index * 2L + 1L) * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, im);
         }
     }
 
@@ -951,7 +951,7 @@ public final class NativeDataAccess {
         } else {
             long address = ((NativeMirror) vector.getNativeMirror()).dataAddress;
             assert address != 0;
-            UnsafeAdapter.UNSAFE.putDouble(address + index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
+            UnsafeAdapter.UNSAFE.putDouble(address + (long) index * Unsafe.ARRAY_DOUBLE_INDEX_SCALE, value);
         }
     }
 
@@ -1267,7 +1267,7 @@ public final class NativeDataAccess {
         long dataAddress = allocateNativeMemory(size);
         long ptr = dataAddress + length * Long.BYTES; // start of the actual character data
         for (int i = 0; i < length; i++) {
-            UnsafeAdapter.UNSAFE.putLong(dataAddress + i * 8, ptr);
+            UnsafeAdapter.UNSAFE.putLong(dataAddress + i * 8L, ptr);
             UnsafeAdapter.UNSAFE.copyMemory(bytes[i], Unsafe.ARRAY_BYTE_BASE_OFFSET, null, ptr, bytes[i].length);
             ptr += bytes[i].length;
             UnsafeAdapter.UNSAFE.putByte(ptr++, (byte) 0);
@@ -1282,7 +1282,7 @@ public final class NativeDataAccess {
         try {
             String[] data = new String[length];
             for (int i = 0; i < length; i++) {
-                long ptr = UnsafeAdapter.UNSAFE.getLong(address + i * 8);
+                long ptr = UnsafeAdapter.UNSAFE.getLong(address + i * 8L);
                 data[i] = readNativeString(ptr);
             }
             return data;
